Moved primary button to buttons folder
